Add a transfer bar and panel to the builder, with a Suggest-transfers search
The session now keeps a baseline (Base/BaseEvent) alongside the imported
squad, so the builder can diff what the reader has changed against what they
imported and price it the way FPL actually would:

- session.Base/BaseEvent, written by importTeam and withheld on a free-hit
  week (the fifteen goes back after the deadline, so it isn't a baseline).
- fpl.HitCost = 4, moved beside MaxBankedTransfers/UnlimitedTransfers;
  internal/backtest.HitCost is now an alias for it.
- viewmodel.State gains an optional Transfers block (free transfers, the
  diff, hits, cost, and why the count might be withheld) and a Move type
  shared with the new suggestion endpoint.
- buildTransferBoard is generalised off an explicit entryID/squad/sell
  triple instead of cfg.EntryID, so the same search serves `armband
  transfers` and a reader's own session without a second implementation.
  A new GET /api/transfers answers with a plan the reader can preview and
  apply, priced at market and respecting their locks/leave-outs.
- The client gets a transfer bar and panel (#transferbar/#transferpanel),
  a Suggest transfers flow with a two-step apply, and a "Put my FPL
  fifteen back" control that reuses the existing import flow rather than
  adding a second reset button. The empty "Your instructions" panel is
  replaced with hiding the whole panel when there is nothing to show.
  A new PLANNING_ELEMENTS/showPlanningSurface(on) switch names the whole
  planning surface in one place for the historical-gameweek work to build
  on next.

Three new golden fixtures cover the transfer panel's states; the existing
pitch-view goldens move with the "Your instructions" change and
/armband-team's goldens are untouched.

+++ internal/backtest/simulate_test.go
@@ -328,7 +328,7 @@ func TestFreeTransfersAreNotFree(t *testing.T) {
 	// a paid transfer stays strictly higher than for a free one.
 	if cost >= HitCost {
 		t.Errorf("a free transfer is charged %.1f against a hit's %.1f; it must be cheaper, "+
-			"or there is never a reason to prefer the free one", cost, HitCost)
+			"or there is never a reason to prefer the free one", cost, float64(HitCost))
 	}
 }
 
@@ -344,7 +344,7 @@ func TestChargingFourForAFreeTransferIsTooMuch(t *testing.T) {
 	if config.Default().Review.FreeTransferValue >= HitCost {
 		t.Errorf("free transfer valued at %.1f, at or above a hit's %.1f — measured at 4 the "+
 			"replay scored 2073 against 2091 for charging nothing",
-			config.Default().Review.FreeTransferValue, HitCost)
+			config.Default().Review.FreeTransferValue, float64(HitCost))
 	}
 	if v := config.Default().Review.FreeTransferValue; v <= 0 {
 		t.Errorf("free transfer valued at %.1f; zero reproduces the churning policy", v)
